Move modified t-digest code into the repo
This change adds in the modified t-digest library crate under a
crates subdirectory in the analytics repository.  It then updates
the toml file to use this local library.

There are also some comments added to the tdigest.rs file within
the extension in an attempt to improve the readability of this file.